Update judges and testing utilities
