+++ setup.py
@@ -87,7 +87,7 @@ EXTRAS = {
     "diffusers": ["diffusers>=0.18.0"],
     # liger-kernel depends on triton, which is only available on Linux https://github.com/triton-lang/triton#compatibility
     "liger": ["liger-kernel>=0.4.0; sys_platform != 'win32'"],
-    "llm_judge": ["openai>=1.23.2", "llm-blender>=0.0.2"],
+    "judges": ["openai>=1.23.2", "llm-blender>=0.0.2"],
     "peft": ["peft>=0.8.0"],
     "quantization": ["bitsandbytes"],
     "scikit": ["scikit-learn"],
