+++ tests/test_xpo_trainer.py
@@ -20,7 +20,9 @@ from transformers import AutoModelForCausalLM, AutoModelForSequenceClassificatio
 from transformers.testing_utils import require_peft
 from transformers.utils import is_peft_available
 
-from trl import RandomPairwiseJudge, XPOConfig, XPOTrainer, is_llm_blender_available
+from trl import XPOConfig, XPOTrainer, is_llm_blender_available
+
+from .testing_utils import RandomPairwiseJudge
 
 
 if is_peft_available():
