+++ trl/__init__.py
@@ -50,8 +50,8 @@ _import_structure = {
         "AlignPropConfig",
         "AlignPropTrainer",
         "AllTrueJudge",
-        "BaseJudge",
         "BaseBinaryJudge",
+        "BaseJudge",
         "BasePairwiseJudge",
         "BaseRankJudge",
         "BCOConfig",
@@ -81,9 +81,6 @@ _import_structure = {
         "PairRMJudge",
         "PPOConfig",
         "PPOTrainer",
-        "RandomBinaryJudge",
-        "RandomPairwiseJudge",
-        "RandomRankJudge",
         "RewardConfig",
         "RewardTrainer",
         "RLOOConfig",
@@ -173,9 +170,6 @@ if TYPE_CHECKING:
         PairRMJudge,
         PPOConfig,
         PPOTrainer,
-        RandomBinaryJudge,
-        RandomPairwiseJudge,
-        RandomRankJudge,
         RewardConfig,
         RewardTrainer,
         RLOOConfig,
